Add web server components
Added in the start of the web server components for the Raspberry Pi server.
myHttp.py, test_suite.py, and webserver.py are being adapted from an earlier
assignment. Still trying to decide if better to call the .cgi and .py robot
scripts directly from html buttons or using a RESTful API protocol.