Fix API call paths
Apparently the string method .find() does not work like I thought it did. I know
it returns the index of the found string or -1 if not found, so I thought using
it as its own boolean would work, but it seems to constantly return True even
when it shouldn't (eg, '<html>' is found nowhere in roboStyles.css, but that
always returned True, even when set to .find('<html>') > 0, so I give up trying
to use it. I have switched to spliting the API call string and just doing a
string comparison on the last token instead.

There's some question as to whether os.system() will call the scripts as indended,
but I can't seemt to test that on my Windows machine, so bumping it to roboberry.