Update webserver for new cgi scripts
Updated the doPOST command in webserver.py to account for the change in cgi
scripts. Also added 'http://' to the WEB_SERVER variable in the frontend
JavaScript because of CORS domain issues without it. Added onclick() functionality
to index.html, even  though I'm not quite at the point of using it yet.